chore: move settings module out of core
