+++ zammad-ai/app/settings/__init__.py
@@ -1,20 +1,22 @@
 from .frontend import FrontendSettings
 from .genai import GenAISettings
+from .kafka import KafkaSettings
 from .qdrant import QdrantSettings
 from .settings import ZammadAISettings, get_settings
 from .triage import TriageSettings
 from .usecase import UseCaseSettings
 from .zammad import BaseZammadSettings, ZammadAPISettings, ZammadEAISettings
 
-__all__ = [
-    "ZammadAISettings",
-    "get_settings",
-    "TriageSettings",
+__all__: list[str] = [
+    "BaseZammadSettings",
     "FrontendSettings",
     "GenAISettings",
+    "get_settings",
+    "KafkaSettings",
     "QdrantSettings",
-    "BaseZammadSettings",
+    "TriageSettings",
+    "UseCaseSettings",
+    "ZammadAISettings",
     "ZammadAPISettings",
     "ZammadEAISettings",
-    "UseCaseSettings",
 ]
