+++ zammad-ai/app/settings/frontend.py
