+++ zammad-ai/app/settings/genai.py
