+++ zammad-ai/app/settings/kafka.py
