+++ zammad-ai/app/settings/qdrant.py
