+++ zammad-ai/app/settings/zammad.py
