+++ zammad-ai/main.py
@@ -11,7 +11,7 @@ from logging import Logger
 import uvicorn
 
 from app.api.backend import backend
-from app.core.settings import ZammadAISettings, get_settings
+from app.settings import ZammadAISettings, get_settings
 from app.utils.logging import get_log_config, getLogger
 
 if __name__ == "__main__":
